Removed vid wraparound logic, moved privacyBucketsOverlap() to PrivacyBucketGroup class, added unit test
