+++ src/main/kotlin/org/wfanet/measurement/eventdataprovider/privacybudgetmanagement/InMemoryBackingStore.kt
@@ -32,64 +32,22 @@ class InMemoryBackingStore : PrivacyBudgetLedgerBackingStore {
 
 class InMemoryBackingStoreTransactionContext(
   val ledger: MutableList<PrivacyBudgetLedgerEntry>,
-  override val transactionId: Long
+  override val transactionId: Long,
 ) : PrivacyBudgetLedgerTransactionContext {
 
   var transactionLedger = ledger.toMutableList()
 
   override fun findIntersectingLedgerEntries(
-    privacyBucketGroup: PrivacyBucketGroup
+    privacyBucketGroup: PrivacyBucketGroup,
   ): List<PrivacyBudgetLedgerEntry> {
     return transactionLedger.filter {
-      privacyBucketsOverlap(privacyBucketGroup, it.privacyBucketGroup)
+      privacyBucketGroup.privacyBucketsOverlap(privacyBucketGroup, it.privacyBucketGroup)
     }
-  }
-
-  fun privacyBucketsOverlap(
-    bucketGroup1: PrivacyBucketGroup,
-    bucketGroup2: PrivacyBucketGroup
-  ): Boolean {
-    if (bucketGroup1.measurementConsumerId != bucketGroup2.measurementConsumerId) {
-      return false
-    }
-    if (bucketGroup2.endingDate.isBefore(bucketGroup1.startingDate) ||
-        bucketGroup1.endingDate.isBefore(bucketGroup2.startingDate)
-    ) {
-      return false
-    }
-    if (bucketGroup1.ageGroup != bucketGroup2.ageGroup) {
-      return false
-    }
-    if (bucketGroup1.gender != bucketGroup2.gender) {
-      return false
-    }
-
-    // The complexity of the following code illustrates why allowing
-    // VID sampling buckets to wrap around was a poor design choice.
-    val vidSampleEnd1 = bucketGroup1.vidSampleStart + bucketGroup1.vidSampleWidth
-    val vidSampleEnd2 = bucketGroup2.vidSampleStart + bucketGroup2.vidSampleWidth
-    if (vidSampleEnd1 > 1.0 && vidSampleEnd2 > 1.0) {
-      // Both intervals contain 0
-      return true
-    }
-
-    if (vidSampleEnd1 > 1.0) {
-      return !((vidSampleEnd1 - 1.0 < bucketGroup2.vidSampleStart) &&
-        (vidSampleEnd2 < bucketGroup1.vidSampleStart))
-    }
-
-    if (vidSampleEnd2 > 1.0) {
-      return !((vidSampleEnd2 - 1.0 < bucketGroup1.vidSampleStart) &&
-        (vidSampleEnd1 < bucketGroup2.vidSampleStart))
-    }
-
-    return (bucketGroup1.vidSampleStart <= vidSampleEnd2) &&
-      (bucketGroup2.vidSampleStart <= vidSampleEnd1)
   }
 
   override fun addLedgerEntry(
     privacyBucketGroup: PrivacyBucketGroup,
-    privacyCharge: PrivacyCharge
+    privacyCharge: PrivacyCharge,
   ) {
     val ledgerEntry =
       PrivacyBudgetLedgerEntry(
